Update dependency of data-model to 0.2.17

+++ src/main/java/org/openlogisticsfoundation/ecmr/domain/beans/ItemBean.java
@@ -21,5 +21,5 @@ class ItemBean {
     private String logisticsPackageType;
     private String transportCargoIdentification;
     private Integer supplyChainConsignmentItemGrossWeight;
-    private Float supplyChainConsignmentItemGrossVolume;
+    private Integer supplyChainConsignmentItemGrossVolume;
 }
